Add a vAccel virtio device
VACCEL_GEN_OP is a generic operation where the frontend has a
prototype of the form:

    function(output, input, output_len, input_len);

in order to serve as a generic abstraction for any kind of
execution.

The form of input and output parameters is completely agnostic
to the framework. vAccel just forwards a call to a specific
function implemented as a plugin. The actual arguments are unpacked
in the plugin and packed in a small wrapper template in the
frontend (not included in this patchset). For clarity, the wrapper
template should be available as an example to help individuals port
their functions to vAccel.

Properly parse genop arguments.

Iterate over the supplied arguments and generate u32, *u8
tuples to be fed to the relevant vaccel-bindings operation.

mem.read is now zero-copy (we only pass the pointer to the
memory space residing in the guest's kernel address space).

compatible with vaccel-bindings branch feat_genop
commit: 165cc0f

Signed-off-by: Anastassios Nanos <ananos@nubificus.co.uk>
Signed-off-by: Babis Chalios <bchalios@nubificus.co.uk>