Fix firecracker to work with zero-copy virtio
The virtio-accel module now supports zero-copy data mapping (it does not
perform copies of the user's arguments during ioctls). Apart from the
potential performance improvement, this avoids `kmallocs` increasing the
size of data buffers we can handle (big kmallocs might fail). The old
way of doing things is still there, choosing the way of handling user
arguments is a compilation option for the kernel module. Firecracker
does not yet support zero-copy, but it needs to handle two extra
descriptors that the kernel module sends us over. That is what this
commit does. In the future we will support the zero-copy version as
well.

Signed-off-by: Babis Chalios <mail@bchalios.io>