Use only vaccel_genop calls to vAccelRT
vAccelRT provides now a vaccel_genop call which operates as a dispatcher
to an arbitrary vAccel API call. It receives a set of read-only and
read-write arguments to pass to the call and the op code of the API
function to be called (as the first read-only argument).

We use this call to replace all individual calls (such as
vaccel_image_classify), to avoid frequent changes to the Firecracker
implementation. Adding a new call to the vAccel API will not require
modifications to Firecracker from now on (hopefully).

Signed-off-by: Babis Chalios <mail@bchalios.io>