dockerfile: we do not need to clone vaccelrt
vAccel bindings crate now builds the correct version of the runtime
itself. As a result, we can remove the step that clones and builds the
runtime inside the container image. We need to keep our own flavour of
the docker image, since we need the llvm packages to build the runtime
anyway (at build time).

Signed-off-by: Babis Chalios <mail@bchalios.io>