Implement `Send` for `vaccel_session`
`vaccel_session`, as exposed by the vaccel bindings crate, includes a
`void *` field now, which renders it non-`Send`. This issue needs to be
fixed in the bindings crate, but until then we add an empty
implementation of `Send` for `vaccel_session`.

Moreover, we change the container of sessions in the vaccel device to
keep `Box`es of `vaccel_session`.

Signed-off-by: Babis Chalios <mail@bchalios.io>